Rename sqlalchemy module to sqlalchemy_utils to avoid relative self import error on Python 2.
